+++ backend/libraries/ledger_utils/src/nns.rs
@@ -29,7 +29,7 @@ pub async fn process_transaction(
 
     match icp_ledger_canister_c2c_client::transfer(transaction.ledger, &transfer_args).await {
         Ok(Ok(block_index)) => Ok(CompletedCryptoTransaction::NNS(types::nns::CompletedCryptoTransaction {
-            // ledger: transaction.ledger,
+            ledger: transaction.ledger,
             token: transaction.token,
             amount: transaction.amount,
             fee,
@@ -51,7 +51,7 @@ pub async fn process_transaction(
     }
     .map_err(|error| {
         FailedCryptoTransaction::NNS(types::nns::FailedCryptoTransaction {
-            // ledger: transaction.ledger,
+            ledger: transaction.ledger,
             token: transaction.token,
             amount: transaction.amount,
             fee,