+++ backend/libraries/ledger_utils/src/sns.rs
@@ -20,7 +20,7 @@ pub async fn process_transaction(
 
     match icrc1_ledger_canister_c2c_client::icrc1_transfer(transaction.ledger, &args).await {
         Ok(Ok(block_index)) => Ok(CompletedCryptoTransaction::SNS(types::sns::CompletedCryptoTransaction {
-            // ledger: transaction.ledger,
+            ledger: transaction.ledger,
             token: transaction.token,
             amount: transaction.amount,
             fee: transaction.fee,
@@ -42,7 +42,7 @@ pub async fn process_transaction(
     }
     .map_err(|error| {
         FailedCryptoTransaction::SNS(types::sns::FailedCryptoTransaction {
-            // ledger: transaction.ledger,
+            ledger: transaction.ledger,
             token: transaction.token,
             amount: transaction.amount,
             fee: transaction.fee,