+++ backend/libraries/types/src/cryptocurrency.rs
@@ -225,9 +225,9 @@ pub mod nns {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "CompletedCryptoTransactionPrevious")]
+    #[serde(from = "CompletedCryptoTransactionPrevious")]
     pub struct CompletedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: Tokens,
         pub fee: Tokens,
@@ -240,9 +240,9 @@ pub mod nns {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "FailedCryptoTransactionPrevious")]
+    #[serde(from = "FailedCryptoTransactionPrevious")]
     pub struct FailedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: Tokens,
         pub fee: Tokens,
@@ -264,31 +264,31 @@ pub mod nns {
         pub created: TimestampNanos,
     }
 
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct CompletedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: Tokens,
-    //     pub fee: Tokens,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub memo: Memo,
-    //     pub created: TimestampNanos,
-    //     pub transaction_hash: TransactionHash,
-    //     pub block_index: BlockIndex,
-    // }
-    //
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct FailedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: Tokens,
-    //     pub fee: Tokens,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub memo: Memo,
-    //     pub created: TimestampNanos,
-    //     pub transaction_hash: TransactionHash,
-    //     pub error_message: String,
-    // }
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct CompletedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: Tokens,
+        pub fee: Tokens,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub memo: Memo,
+        pub created: TimestampNanos,
+        pub transaction_hash: TransactionHash,
+        pub block_index: BlockIndex,
+    }
+
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct FailedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: Tokens,
+        pub fee: Tokens,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub memo: Memo,
+        pub created: TimestampNanos,
+        pub transaction_hash: TransactionHash,
+        pub error_message: String,
+    }
 
     impl From<PendingCryptoTransactionPrevious> for PendingCryptoTransaction {
         fn from(value: PendingCryptoTransactionPrevious) -> Self {
@@ -304,39 +304,39 @@ pub mod nns {
         }
     }
 
-    // impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
-    //     fn from(value: CompletedCryptoTransactionPrevious) -> Self {
-    //         CompletedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             from: value.from,
-    //             to: value.to,
-    //             fee: value.fee,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             transaction_hash: value.transaction_hash,
-    //             block_index: value.block_index,
-    //         }
-    //     }
-    // }
-    //
-    // impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
-    //     fn from(value: FailedCryptoTransactionPrevious) -> Self {
-    //         FailedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             fee: value.fee,
-    //             from: value.from,
-    //             to: value.to,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             transaction_hash: value.transaction_hash,
-    //             error_message: value.error_message,
-    //         }
-    //     }
-    // }
+    impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
+        fn from(value: CompletedCryptoTransactionPrevious) -> Self {
+            CompletedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                from: value.from,
+                to: value.to,
+                fee: value.fee,
+                memo: value.memo,
+                created: value.created,
+                transaction_hash: value.transaction_hash,
+                block_index: value.block_index,
+            }
+        }
+    }
+
+    impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
+        fn from(value: FailedCryptoTransactionPrevious) -> Self {
+            FailedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                fee: value.fee,
+                from: value.from,
+                to: value.to,
+                memo: value.memo,
+                created: value.created,
+                transaction_hash: value.transaction_hash,
+                error_message: value.error_message,
+            }
+        }
+    }
 }
 
 pub mod sns {
@@ -364,9 +364,9 @@ pub mod sns {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "CompletedCryptoTransactionPrevious")]
+    #[serde(from = "CompletedCryptoTransactionPrevious")]
     pub struct CompletedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: Tokens,
         pub fee: Tokens,
@@ -380,9 +380,9 @@ pub mod sns {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "FailedCryptoTransactionPrevious")]
+    #[serde(from = "FailedCryptoTransactionPrevious")]
     pub struct FailedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: Tokens,
         pub fee: Tokens,
@@ -405,31 +405,31 @@ pub mod sns {
         pub created: TimestampNanos,
     }
 
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct CompletedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: Tokens,
-    //     pub fee: Tokens,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub memo: Option<Memo>,
-    //     pub created: TimestampNanos,
-    //     pub transaction_hash: TransactionHash,
-    //     pub block_index: BlockIndex,
-    // }
-    //
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct FailedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: Tokens,
-    //     pub fee: Tokens,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub memo: Option<Memo>,
-    //     pub created: TimestampNanos,
-    //     pub transaction_hash: TransactionHash,
-    //     pub error_message: String,
-    // }
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct CompletedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: Tokens,
+        pub fee: Tokens,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub memo: Option<Memo>,
+        pub created: TimestampNanos,
+        pub transaction_hash: TransactionHash,
+        pub block_index: BlockIndex,
+    }
+
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct FailedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: Tokens,
+        pub fee: Tokens,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub memo: Option<Memo>,
+        pub created: TimestampNanos,
+        pub transaction_hash: TransactionHash,
+        pub error_message: String,
+    }
 
     impl From<PendingCryptoTransactionPrevious> for PendingCryptoTransaction {
         fn from(value: PendingCryptoTransactionPrevious) -> Self {
@@ -445,39 +445,39 @@ pub mod sns {
         }
     }
 
-    // impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
-    //     fn from(value: CompletedCryptoTransactionPrevious) -> Self {
-    //         CompletedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             from: value.from,
-    //             to: value.to,
-    //             fee: value.fee,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             transaction_hash: value.transaction_hash,
-    //             block_index: value.block_index,
-    //         }
-    //     }
-    // }
-    //
-    // impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
-    //     fn from(value: FailedCryptoTransactionPrevious) -> Self {
-    //         FailedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             fee: value.fee,
-    //             from: value.from,
-    //             to: value.to,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             transaction_hash: value.transaction_hash,
-    //             error_message: value.error_message,
-    //         }
-    //     }
-    // }
+    impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
+        fn from(value: CompletedCryptoTransactionPrevious) -> Self {
+            CompletedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                from: value.from,
+                to: value.to,
+                fee: value.fee,
+                memo: value.memo,
+                created: value.created,
+                transaction_hash: value.transaction_hash,
+                block_index: value.block_index,
+            }
+        }
+    }
+
+    impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
+        fn from(value: FailedCryptoTransactionPrevious) -> Self {
+            FailedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                fee: value.fee,
+                from: value.from,
+                to: value.to,
+                memo: value.memo,
+                created: value.created,
+                transaction_hash: value.transaction_hash,
+                error_message: value.error_message,
+            }
+        }
+    }
 }
 
 pub mod icrc1 {
@@ -579,9 +579,9 @@ pub mod icrc1 {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "CompletedCryptoTransactionPrevious")]
+    #[serde(from = "CompletedCryptoTransactionPrevious")]
     pub struct CompletedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: u128,
         pub from: CryptoAccount,
@@ -593,9 +593,9 @@ pub mod icrc1 {
     }
 
     #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // #[serde(from = "FailedCryptoTransactionPrevious")]
+    #[serde(from = "FailedCryptoTransactionPrevious")]
     pub struct FailedCryptoTransaction {
-        // pub ledger: CanisterId,
+        pub ledger: CanisterId,
         pub token: Cryptocurrency,
         pub amount: u128,
         pub fee: u128,
@@ -616,29 +616,29 @@ pub mod icrc1 {
         pub created: TimestampNanos,
     }
 
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct CompletedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: u128,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub fee: u128,
-    //     pub memo: Option<Memo>,
-    //     pub created: TimestampNanos,
-    //     pub block_index: u64,
-    // }
-    //
-    // #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
-    // pub struct FailedCryptoTransactionPrevious {
-    //     pub token: Cryptocurrency,
-    //     pub amount: u128,
-    //     pub fee: u128,
-    //     pub from: CryptoAccount,
-    //     pub to: CryptoAccount,
-    //     pub memo: Option<Memo>,
-    //     pub created: TimestampNanos,
-    //     pub error_message: String,
-    // }
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct CompletedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: u128,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub fee: u128,
+        pub memo: Option<Memo>,
+        pub created: TimestampNanos,
+        pub block_index: u64,
+    }
+
+    #[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+    pub struct FailedCryptoTransactionPrevious {
+        pub token: Cryptocurrency,
+        pub amount: u128,
+        pub fee: u128,
+        pub from: CryptoAccount,
+        pub to: CryptoAccount,
+        pub memo: Option<Memo>,
+        pub created: TimestampNanos,
+        pub error_message: String,
+    }
 
     impl From<PendingCryptoTransactionPrevious> for PendingCryptoTransaction {
         fn from(value: PendingCryptoTransactionPrevious) -> Self {
@@ -654,35 +654,35 @@ pub mod icrc1 {
         }
     }
 
-    // impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
-    //     fn from(value: CompletedCryptoTransactionPrevious) -> Self {
-    //         CompletedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             from: value.from,
-    //             to: value.to,
-    //             fee: value.fee,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             block_index: value.block_index,
-    //         }
-    //     }
-    // }
-    //
-    // impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
-    //     fn from(value: FailedCryptoTransactionPrevious) -> Self {
-    //         FailedCryptoTransaction {
-    //             ledger: value.token.ledger_canister_id(),
-    //             token: value.token,
-    //             amount: value.amount,
-    //             fee: value.fee,
-    //             from: value.from,
-    //             to: value.to,
-    //             memo: value.memo,
-    //             created: value.created,
-    //             error_message: value.error_message,
-    //         }
-    //     }
-    // }
+    impl From<CompletedCryptoTransactionPrevious> for CompletedCryptoTransaction {
+        fn from(value: CompletedCryptoTransactionPrevious) -> Self {
+            CompletedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                from: value.from,
+                to: value.to,
+                fee: value.fee,
+                memo: value.memo,
+                created: value.created,
+                block_index: value.block_index,
+            }
+        }
+    }
+
+    impl From<FailedCryptoTransactionPrevious> for FailedCryptoTransaction {
+        fn from(value: FailedCryptoTransactionPrevious) -> Self {
+            FailedCryptoTransaction {
+                ledger: value.token.ledger_canister_id(),
+                token: value.token,
+                amount: value.amount,
+                fee: value.fee,
+                from: value.from,
+                to: value.to,
+                memo: value.memo,
+                created: value.created,
+                error_message: value.error_message,
+            }
+        }
+    }
 }